Move instance types to subpackages
