+++ src/main/java/no/unit/nva/model/instancetypes/book/BookAnthology.java
@@ -1,31 +1,27 @@
-package no.unit.nva.model.instancetypes;
+package no.unit.nva.model.instancetypes.book;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import no.unit.nva.model.instancetypes.PeerReviewedMonograph;
 import no.unit.nva.model.pages.MonographPages;
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
-public class BookMonograph extends PeerReviewedMonograph {
+public class BookAnthology extends PeerReviewedMonograph {
 
-    public BookMonograph(@JsonProperty("pages") MonographPages pages,
+    public BookAnthology(@JsonProperty("pages") MonographPages pages,
                          @JsonProperty("peerReviewed") boolean peerReviewed) {
         super(pages, peerReviewed);
     }
 
-    private BookMonograph(Builder builder) {
+    private BookAnthology(Builder builder) {
         super(builder.pages, builder.peerReviewed);
     }
 
     public static final class Builder {
-        private boolean peerReviewed;
         private MonographPages pages;
+        private boolean peerReviewed;
 
         public Builder() {
-        }
-
-        public Builder withPeerReviewed(boolean peerReviewed) {
-            this.peerReviewed = peerReviewed;
-            return this;
         }
 
         public Builder withPages(MonographPages pages) {
@@ -33,8 +29,13 @@ public class BookMonograph extends PeerReviewedMonograph {
             return this;
         }
 
-        public BookMonograph build() {
-            return new BookMonograph(this);
+        public Builder withPeerReviewed(boolean peerReviewed) {
+            this.peerReviewed = peerReviewed;
+            return this;
+        }
+
+        public BookAnthology build() {
+            return new BookAnthology(this);
         }
     }
 }
