+++ src/main/java/no/unit/nva/model/instancetypes/book/BookMonograph.java
@@ -1,31 +1,27 @@
-package no.unit.nva.model.instancetypes;
+package no.unit.nva.model.instancetypes.book;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import no.unit.nva.model.instancetypes.PeerReviewedMonograph;
 import no.unit.nva.model.pages.MonographPages;
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
-public class BookAnthology extends PeerReviewedMonograph {
+public class BookMonograph extends PeerReviewedMonograph {
 
-    public BookAnthology(@JsonProperty("pages") MonographPages pages,
+    public BookMonograph(@JsonProperty("pages") MonographPages pages,
                          @JsonProperty("peerReviewed") boolean peerReviewed) {
         super(pages, peerReviewed);
     }
 
-    private BookAnthology(Builder builder) {
+    private BookMonograph(Builder builder) {
         super(builder.pages, builder.peerReviewed);
     }
 
     public static final class Builder {
-        private MonographPages pages;
         private boolean peerReviewed;
+        private MonographPages pages;
 
         public Builder() {
-        }
-
-        public Builder withPages(MonographPages pages) {
-            this.pages = pages;
-            return this;
         }
 
         public Builder withPeerReviewed(boolean peerReviewed) {
@@ -33,8 +29,13 @@ public class BookAnthology extends PeerReviewedMonograph {
             return this;
         }
 
-        public BookAnthology build() {
-            return new BookAnthology(this);
+        public Builder withPages(MonographPages pages) {
+            this.pages = pages;
+            return this;
+        }
+
+        public BookMonograph build() {
+            return new BookMonograph(this);
         }
     }
 }
