+++ src/main/java/no/unit/nva/model/instancetypes/report/ReportWorkingPaper.java
@@ -1,17 +1,18 @@
-package no.unit.nva.model.instancetypes;
+package no.unit.nva.model.instancetypes.report;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import no.unit.nva.model.instancetypes.NonPeerReviewedMonograph;
 import no.unit.nva.model.pages.MonographPages;
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
-public class Report extends NonPeerReviewedMonograph {
+public class ReportWorkingPaper extends NonPeerReviewedMonograph {
 
-    public Report(@JsonProperty("pages") MonographPages pages) {
+    public ReportWorkingPaper(@JsonProperty("pages") MonographPages pages) {
         super(pages);
     }
 
-    private Report(Builder builder) {
+    private ReportWorkingPaper(Builder builder) {
         this(builder.pages);
     }
 
@@ -26,9 +27,8 @@ public class Report extends NonPeerReviewedMonograph {
             return this;
         }
 
-        public Report build() {
-            return new Report(this);
+        public ReportWorkingPaper build() {
+            return new ReportWorkingPaper(this);
         }
     }
-
 }
