+++ src/test/java/no/unit/nva/model/ModelTest.java
@@ -11,21 +11,21 @@ import no.unit.nva.model.exceptions.InvalidIsbnException;
 import no.unit.nva.model.exceptions.InvalidIssnException;
 import no.unit.nva.model.exceptions.InvalidPageRangeException;
 import no.unit.nva.model.exceptions.MalformedContributorException;
-import no.unit.nva.model.instancetypes.BookAnthology;
-import no.unit.nva.model.instancetypes.BookMonograph;
-import no.unit.nva.model.instancetypes.ChapterArticle;
-import no.unit.nva.model.instancetypes.DegreeBachelor;
-import no.unit.nva.model.instancetypes.DegreeMaster;
-import no.unit.nva.model.instancetypes.DegreePhd;
-import no.unit.nva.model.instancetypes.JournalArticle;
-import no.unit.nva.model.instancetypes.JournalLeader;
-import no.unit.nva.model.instancetypes.JournalLetter;
-import no.unit.nva.model.instancetypes.JournalReview;
-import no.unit.nva.model.instancetypes.JournalShortCommunication;
+import no.unit.nva.model.instancetypes.book.BookAnthology;
+import no.unit.nva.model.instancetypes.book.BookMonograph;
+import no.unit.nva.model.instancetypes.chapter.ChapterArticle;
+import no.unit.nva.model.instancetypes.degree.DegreeBachelor;
+import no.unit.nva.model.instancetypes.degree.DegreeMaster;
+import no.unit.nva.model.instancetypes.degree.DegreePhd;
+import no.unit.nva.model.instancetypes.journal.JournalArticle;
+import no.unit.nva.model.instancetypes.journal.JournalLeader;
+import no.unit.nva.model.instancetypes.journal.JournalLetter;
+import no.unit.nva.model.instancetypes.journal.JournalReview;
+import no.unit.nva.model.instancetypes.journal.JournalShortCommunication;
 import no.unit.nva.model.instancetypes.PublicationInstance;
-import no.unit.nva.model.instancetypes.ReportPolicy;
-import no.unit.nva.model.instancetypes.ReportResearch;
-import no.unit.nva.model.instancetypes.ReportWorkingPaper;
+import no.unit.nva.model.instancetypes.report.ReportPolicy;
+import no.unit.nva.model.instancetypes.report.ReportResearch;
+import no.unit.nva.model.instancetypes.report.ReportWorkingPaper;
 import no.unit.nva.model.pages.MonographPages;
 import no.unit.nva.model.pages.Pages;
 import no.unit.nva.model.pages.Range;
