+++ src/test/java/no/unit/nva/model/instancetypes/DegreePhdTest.java
@@ -1,8 +1,8 @@
 package no.unit.nva.model.instancetypes;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import no.unit.nva.model.contexttypes.Degree;
 import no.unit.nva.model.exceptions.InvalidPageRangeException;
+import no.unit.nva.model.instancetypes.degree.DegreePhd;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.function.Executable;
