+++ src/test/java/no/unit/nva/model/util/PublicationGenerator.java
@@ -25,12 +25,12 @@ import no.unit.nva.model.exceptions.InvalidIsbnException;
 import no.unit.nva.model.exceptions.InvalidIssnException;
 import no.unit.nva.model.exceptions.InvalidPageRangeException;
 import no.unit.nva.model.exceptions.MalformedContributorException;
-import no.unit.nva.model.instancetypes.BookAnthology;
-import no.unit.nva.model.instancetypes.JournalArticle;
-import no.unit.nva.model.instancetypes.JournalLeader;
-import no.unit.nva.model.instancetypes.JournalLetter;
-import no.unit.nva.model.instancetypes.JournalReview;
-import no.unit.nva.model.instancetypes.JournalShortCommunication;
+import no.unit.nva.model.instancetypes.book.BookAnthology;
+import no.unit.nva.model.instancetypes.journal.JournalArticle;
+import no.unit.nva.model.instancetypes.journal.JournalLeader;
+import no.unit.nva.model.instancetypes.journal.JournalLetter;
+import no.unit.nva.model.instancetypes.journal.JournalReview;
+import no.unit.nva.model.instancetypes.journal.JournalShortCommunication;
 import no.unit.nva.model.instancetypes.PublicationInstance;
 import no.unit.nva.model.pages.MonographPages;
 import no.unit.nva.model.pages.Range;
